Improve debug window
* Fix scrolling (#4167)
* Open bobs with double click
* Change structs to classes
